Added the solution requirements
I split the requirements into two sections, frontend, and backend. These sections were then split into 'functional requirements', and 'non-functional requirements'. Functional requirements, are those that are necessary, in order for the software to do its job. While the non-functional requirements are their for flair, and usefulness, such as having the software have a quick response time, it is not necessary, but it is very useful and assists in improving customer experience.