Added data to the database
I tweaked the database creation code slightly, as SQLite doesn't have a 'DATE' datatype, but instead dates can be stored as strings or integers. I also created a new SQLite script that populates the database, with an insane amount of data. I used ChatGPT to aide in the creation of the data, so the prompt for that has been referenced